+++ src/screens/listings/AllListings.js
@@ -9,11 +9,11 @@ import {
   Modal,
   TouchableWithoutFeedback,
 } from "react-native";
-import firebase from "../../database/Firebase";
+import firebase from "../../../database/Firebase";
 import "firebase/auth";
 import { Ionicons } from "@expo/vector-icons";
 
-import Categories from "../../constants/Categories";
+import Categories from "../../../constants/Categories";
 
 const sortOptions = [
   { key: "dateLatestToOldest", label: "Date Listed: Newest to Oldest" },
